+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Reworked Crafting Monitor GUI (raoulvdberge)
 - Fixed problems relating to Crafting Upgrade (scheduling a task wrongly, blocking other tasks, etc) (raoulvdberge)
 - Interface now supports Crafting Upgrade (raoulvdberge)
+- When shift clicking a recipe in the Crafting Grid, the player's inventory is now leveraged as well (raoulvdberge)
 
 ### 0.9.4
 - Little fixes in German translation (ThexXTURBOXx)